requirements.txt

pandas==2.2.3
numpy==2.1.3
PyYAML==6.0.2
transformers==4.56.0
datasets==2.21.0
accelerate==1.1.1
tiktoken==0.7.0
sentencepiece==0.2.0
huggingface_hub==0.26.2
scikit-learn==1.5.2
matplotlib==3.9.2
fastapi==0.115.5
uvicorn[standard]==0.32.0
prometheus-client==0.21.0
pydantic==2.9.2
pydantic-settings==2.6.1
typer==0.12.5
rich==13.9.4
streamlit==1.40.1
